requirements.txt
discord.py>=2.3.0
requests>=2.31.0
groq>=0.4.0
